+++ bin/rmemo.js
@@ -18,6 +18,7 @@ import { cmdPr } from "../src/cmd/pr.js";
 import { cmdWatch } from "../src/cmd/watch.js";
 import { cmdWs } from "../src/cmd/ws.js";
 import { cmdProfile } from "../src/cmd/profile.js";
+import { cmdSession } from "../src/cmd/session.js";
 import { parseArgs, printHelp } from "../src/lib/args.js";
 import { exitWithError } from "../src/lib/io.js";
 
@@ -82,6 +83,9 @@ try {
       break;
     case "profile":
       await cmdProfile({ rest, flags });
+      break;
+    case "session":
+      await cmdSession({ rest, flags });
       break;
     case "help":
     case undefined: